feat(posts): add new post "Shame Doesn't Change Minds" and update index to include it
The commit adds a new post titled "Shame Doesn't Change Minds" to the posts directory. The post is about how shame is not an effective tool for changing people's minds and how productive conversations and argument-based memes can be more effective. The commit also updates the index to include a link to the new post.